delete code updated

+++ routers/blogs.js
@@ -53,9 +53,11 @@ router.post("/", uploadOptions.single("image"), async (req, res) => {
         if (err) throw err;
 
         for (const file of files) {
-          fs.unlink(path.join(directory, file), (err) => {
-            if (err) throw err;
-          });
+          if (file !== "demo.txt") {
+            fs.unlink(path.join(directory, file), (err) => {
+              if (err) throw err;
+            });
+          }
         }
       });
     }
@@ -100,9 +102,11 @@ router.put("/:id", uploadOptions.single("image"), async (req, res) => {
         if (err) throw err;
 
         for (const file of files) {
-          fs.unlink(path.join(directory, file), (err) => {
-            if (err) throw err;
-          });
+          if (file !== "demo.txt") {
+            fs.unlink(path.join(directory, file), (err) => {
+              if (err) throw err;
+            });
+          }
         }
       });
     }
