+++ routers/categories.js
@@ -60,9 +60,11 @@ router.post("/", uploadOptions.single("image"), async (req, res) => {
       if (err) throw err;
 
       for (const file of files) {
-        fs.unlink(path.join(directory, file), (err) => {
-          if (err) throw err;
-        });
+        if (file !== "demo.txt") {
+          fs.unlink(path.join(directory, file), (err) => {
+            if (err) throw err;
+          });
+        }
       }
     });
   }
@@ -94,9 +96,11 @@ router.put("/:id", uploadOptions.single("image"), async (req, res) => {
         if (err) throw err;
 
         for (const file of files) {
-          fs.unlink(path.join(directory, file), (err) => {
-            if (err) throw err;
-          });
+          if (file !== "demo.txt") {
+            fs.unlink(path.join(directory, file), (err) => {
+              if (err) throw err;
+            });
+          }
         }
       });
     }
