+++ routers/products.js
@@ -96,9 +96,11 @@ router.post(`/`, uploadOptions.single("image"), async (req, res) => {
       if (err) throw err;
 
       for (const file of files) {
-        fs.unlink(path.join(directory, file), (err) => {
-          if (err) throw err;
-        });
+        if (file !== "demo.txt") {
+          fs.unlink(path.join(directory, file), (err) => {
+            if (err) throw err;
+          });
+        }
       }
     });
   }
@@ -144,9 +146,11 @@ router.put("/:id", uploadOptions.single("image"), async (req, res) => {
         if (err) throw err;
 
         for (const file of files) {
-          fs.unlink(path.join(directory, file), (err) => {
-            if (err) throw err;
-          });
+          if (file !== "demo.txt") {
+            fs.unlink(path.join(directory, file), (err) => {
+              if (err) throw err;
+            });
+          }
         }
       });
     }
@@ -175,9 +179,11 @@ router.put("/:id", uploadOptions.single("image"), async (req, res) => {
         if (err) throw err;
 
         for (const file of files) {
-          fs.unlink(path.join(directory, file), (err) => {
-            if (err) throw err;
-          });
+          if (file !== "demo.txt") {
+            fs.unlink(path.join(directory, file), (err) => {
+              if (err) throw err;
+            });
+          }
         }
       });
     }
@@ -272,9 +278,11 @@ router.put(
         if (err) throw err;
 
         for (const file of files) {
-          fs.unlink(path.join(directory, file), (err) => {
-            if (err) throw err;
-          });
+          if (file !== "demo.txt") {
+            fs.unlink(path.join(directory, file), (err) => {
+              if (err) throw err;
+            });
+          }
         }
       });
     }
